requirements.txt
# Core ML & Image Processing
tensorflow
numpy
matplotlib
opencv-python
scikit-image

# API & Web Server
flask
flask-cors
werkzeug
streamlit

# File Handling & Utilities
h5py
requests
pillow  # For image handling

# Testing & Debugging
pytest
